pyproject.toml
[project]
name = "fashion-multi-modal-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.13.0",
    "datasets>=4.8.5",
    "fastapi>=0.136.1",
    "flashrank>=0.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-text-splitters>=1.1.2",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "qdrant-client>=1.17.1",
    "scipy>=1.17.1",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.57.0",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
    "transformers>=5.7.0",
    "unstructured[all-docs]>=0.22.18",
    "uvicorn>=0.46.0",
]
